Added simple level editor, repackaged custom views, split FieldView to two entities
