+++ Labyrinth/app/src/main/AndroidManifest.xml
@@ -9,24 +9,25 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
+
         <activity
             android:name=".MainActivity"
-            android:screenOrientation="portrait"></activity>
-        <activity android:name=".LogActivity"></activity>
+            android:screenOrientation="portrait"> </activity>
+
+        <activity android:name=".LogActivity"> </activity>
+
         <activity android:name=".StartActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".InitActivity">
-            <!--<intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>-->
-        </activity>
+        <activity android:name=".InitActivity"> </activity>
+
+        <activity
+            android:name="EditorActivity"
+            android:screenOrientation="portrait"> </activity>
     </application>
 
 </manifest>
