+++ Labyrinth/app/src/main/java/ru/spbau/labyrinth/customviews/DirectionChooseView.java
@@ -1,4 +1,4 @@
-package ru.spbau.labyrinth;
+package ru.spbau.labyrinth.customviews;
 
 import android.content.Context;
 import android.graphics.Canvas;
