+++ Labyrinth/app/src/main/res/layout/activity_log.xml
@@ -11,11 +11,11 @@
         android:layout_width="900dp"
         android:layout_height="900dp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.494"
+        app:layout_constraintHorizontal_bias="0.501"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.39"
+        app:layout_constraintVertical_bias="0.443"
         app:srcCompat="@drawable/labyrinth" />
 
     <TextView
